added utils for generalization
