+++ utils/items.py
@@ -1,13 +1,21 @@
-from utils.items import EdgeItem, VertexItem
+"""
+Scrapy Items
+---
+BIDS Institutional Ecosystem Mapping
+
+These are ScraPy items emitted by the scraper. ScrapyHub only saves these items.
+"""
+
+import scrapy
 
 
-class LinkItem(VertexItem):
+class EdgeItem(scrapy.Item):
     base_url = scrapy.Field()
     url = scrapy.Field()
     link = scrapy.Field()
     timestamp = scrapy.Field()
 
-class TextItem(VertexItem):
+class VertexItem(scrapy.Item):
     base_url = scrapy.Field()
     url = scrapy.Field()
     text = scrapy.Field()
